replace inplaceabnsync to typical abn for inference in non-cuda machines
