+++ simple_extractor.py
@@ -56,7 +56,7 @@ def get_arguments():
 
     parser.add_argument("--dataset", type=str, default='lip', choices=['lip', 'atr', 'pascal'])
     parser.add_argument("--model-restore", type=str, default='', help="restore pretrained model parameters.")
-    parser.add_argument("--gpu", type=str, default='0', help="choose gpu device.")
+    parser.add_argument("--gpu", type=str, default='None', help="choose gpu device.")
     parser.add_argument("--input-dir", type=str, default='', help="path of input image folder.")
     parser.add_argument("--output-dir", type=str, default='', help="path of output image folder.")
     parser.add_argument("--logits", action='store_true', default=False, help="whether to save the logits.")
@@ -90,10 +90,10 @@ def get_palette(num_cls):
 
 def main():
     args = get_arguments()
-
-    gpus = [int(i) for i in args.gpu.split(',')]
-    assert len(gpus) == 1
-    if not args.gpu == 'None':
+    if args.gpu != 'None':
+        gpus = [int(i) for i in args.gpu.split(',')]
+        assert len(gpus) == 1
+    if args.gpu != 'None':
         os.environ["CUDA_VISIBLE_DEVICES"] = args.gpu
 
     num_classes = dataset_settings[args.dataset]['num_classes']
@@ -110,7 +110,8 @@ def main():
         name = k[7:]  # remove `module.`
         new_state_dict[name] = v
     model.load_state_dict(new_state_dict)
-    model.cuda()
+    if args.gpu != 'None':
+        model.cuda()
     model.eval()
 
     transform = transforms.Compose([
@@ -133,7 +134,10 @@ def main():
             w = meta['width'].numpy()[0]
             h = meta['height'].numpy()[0]
 
-            output = model(image.cuda())
+            if args.gpu != 'None':
+                output = model(image.cuda())
+            else:
+                output = model(image)
             upsample = torch.nn.Upsample(size=input_size, mode='bilinear', align_corners=True)
             upsample_output = upsample(output[0][-1][0].unsqueeze(0))
             upsample_output = upsample_output.squeeze()
